# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  DTO.PaginateResponse:
    properties:
      count:
        type: integer
      data: {}
    type: object
...
  DTO.QuestionGroupResponse:
    properties:
      SpeakingTopic:
        $ref: '#/definitions/models.QuestionGroupSpeaking'
      componentType:
        type: string
      description:
        type: string
      id:
        type: string
      identifyingInformation:
        items:
          $ref: '#/definitions/models.QuestionGroupIdentifyingInformation'
        type: array
      map:
        $ref: '#/definitions/models.QuestionGroupMap'
      multipleChoice:
        items:
          $ref: '#/definitions/models.QuestionGroupMultipleChoice'
        type: array
      noteCompletion:
        $ref: '#/definitions/models.QuestionGroupNoteCompletion'
      q_end:
        type: integer
      q_start:
        type: integer
      questions:
        items:
          $ref: '#/definitions/DTO.QuestionWithoutAnswerResponse'
        type: array
      title:
        type: string
      writingTopic:
        $ref: '#/definitions/models.QuestionGroupWriting'
    type: object
...
  DTO.QuestionStatisticResponse:
    properties:
      id:
        type: string
      questionType:
        type: string
      t_answer:
        type: integer
      t_correctAnswer:
        type: integer
    type: object
...
  DTO.QuestionWithUserAnswerResponse:
    properties:
      correctAnswer:
        type: string
      id:
        type: string
      q_number:
        type: integer
      statistic:
        $ref: '#/definitions/DTO.QuestionStatisticResponse'
      userAnswer:
        $ref: '#/definitions/DTO.UserAnswerResponse'
    type: object
  DTO.QuestionWithoutAnswerResponse:
    properties:
      id:
        type: string
      q_number:
        type: integer
    type: object
  DTO.SectionCreateRequest:
    properties:
      componentType:
        type: integer
      description:
        type: string
      q_end:
        type: integer
      q_start:
        type: integer
      sectionListening:
        $ref: '#/definitions/models.SectionListening'
      sectionReading:
        $ref: '#/definitions/models.SectionReading'
      sectionSpeaking:
        $ref: '#/definitions/models.SectionSpeaking'
      sectionWriting:
        $ref: '#/definitions/models.SectionWriting'
      testId:
        type: string
      title:
        type: string
      totalSectionTime:
        type: integer
    type: object
  DTO.SectionCreateResponse:
    properties:
      componentType:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      sectionListening:
        $ref: '#/definitions/models.SectionListening'
      sectionReading:
        $ref: '#/definitions/models.SectionReading'
      sectionSpeaking:
        $ref: '#/definitions/models.SectionSpeaking'
      sectionWriting:
        $ref: '#/definitions/models.SectionWriting'
      title:
        type: string
    type: object
  DTO.SectionResponse:
    properties:
      componentType:
        type: string
      description:
        type: string
      id:
        type: string
      q_end:
        type: integer
      q_start:
        type: integer
      questionGroups:
        items:
          $ref: '#/definitions/DTO.QuestionGroupResponse'
        type: array
      sectionListening:
        $ref: '#/definitions/models.SectionListening'
      sectionReading:
        $ref: '#/definitions/models.SectionReading'
      sectionSpeaking:
        $ref: '#/definitions/models.SectionSpeaking'
      sectionWriting:
        $ref: '#/definitions/models.SectionWriting'
      title:
        type: string
      totalSectionTime:
        type: integer
    type: object
  DTO.SectionUpdateRequest:
    properties:
      componentType:
        type: integer
      description:
        type: string
      id:
        type: string
      q_end:
        type: integer
      q_start:
        type: integer
      sectionListening:
        $ref: '#/definitions/models.SectionListening'
      sectionReading:
        $ref: '#/definitions/models.SectionReading'
      sectionSpeaking:
        $ref: '#/definitions/models.SectionSpeaking'
      sectionWriting:
        $ref: '#/definitions/models.SectionWriting'
      title:
        type: string
      totalSectionTime:
        type: integer
    type: object
  DTO.SectionUpdateResponse:
    properties:
      componentType:
        type: string
      description:
        type: string
      id:
        type: string
      sectionListening:
        $ref: '#/definitions/models.SectionListening'
      sectionReading:
        $ref: '#/definitions/models.SectionReading'
      sectionSpeaking:
        $ref: '#/definitions/models.SectionSpeaking'
      sectionWriting:
        $ref: '#/definitions/models.SectionWriting'
      title:
        type: string
      updateAt:
        type: string
    type: object
...
  DTO.UserAnswerResponse:
    properties:
      answer:
        type: string
      id:
        type: string
      markerComment:
        type: string
      markerId:
        type: string
      markerScore:
        type: number
      questionId:
        type: string
      userTestId:
        type: string
    type: object
...
  DTO.UserSignUpOrLoginUserResponse:
    properties:
      alias:
        type: string
      email:
        type: string
      subscriptionExpirationMs:
        type: integer
      subscriptionState:
        type: string
      userId:
        type: string
    type: object
  DTO.UserSkillsAvgStat:
    properties:
      listeningOverall:
        type: number
      readingOverall:
        type: number
      speakingOverall:
        type: number
      writingOverall:
        type: number
    type: object
  DTO.UserTestCreateRequest:
    properties:
      testDate:
        type: integer
      testId:
        type: string
      testType:
        type: integer
      userId:
        type: string
    type: object
  DTO.UserTestCreateResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      testDate:
        type: string
      testDuration:
        type: integer
      testId:
        type: string
      testType:
        type: string
      userId:
        type: string
    type: object
  DTO.UserTestResponse:
    properties:
      id:
        type: string
      testDate:
        type: string
      testDuration:
        type: integer
      testId:
        type: string
      testType:
        type: string
      userAnswers:
        items:
          $ref: '#/definitions/DTO.UserAnswerResponse'
        type: array
      userId:
        type: string
    type: object
  DTO.UserTestResultResponse:
    properties:
      answerSheet:
        items:
          $ref: '#/definitions/DTO.QuestionWithUserAnswerResponse'
        type: array
      id:
        type: string
      listeningScore:
        type: number
      overallScore:
        type: number
      readingScore:
        type: number
      speakingScore:
        type: number
      stat:
        $ref: '#/definitions/DTO.UserTestStat'
      testDate:
        type: string
      testDuration:
        type: integer
      testId:
        type: string
      testType:
        type: string
      userId:
        type: string
      writingScore:
        type: number
    type: object
  DTO.UserTestStat:
    properties:
      average:
        $ref: '#/definitions/DTO.UserSkillsAvgStat'
      averageTotalSkill:
        $ref: '#/definitions/DTO.UserSkillsAvgStat'
      skill:
        $ref: '#/definitions/models.TestSkillsStat'
    type: object
...
  models.QuestionGroupIdentifyingInformation:
    properties:
      q_number:
        type: integer
      title:
        type: string
    type: object
  models.QuestionGroupMap:
    properties:
      items:
        items:
          $ref: '#/definitions/models.QuestionGroupMapItem'
        type: array
      mapUrl:
        type: string
    type: object
  models.QuestionGroupMapItem:
    properties:
      q_number:
        type: integer
      title:
        type: string
    type: object
  models.QuestionGroupMultipleChoice:
    properties:
      answers:
        additionalProperties:
          type: string
        type: object
      q_number:
        type: integer
      title:
        type: string
    type: object
  models.QuestionGroupNoteCompletion:
    properties:
      note:
        type: string
    type: object
  models.QuestionGroupSpeaking:
    properties:
      q_number:
        type: integer
      topic:
        type: string
    type: object
  models.QuestionGroupWriting:
    properties:
      q_number:
        type: integer
      topic:
        type: string
    type: object
  models.SectionListening:
    properties:
      mediaLengthMin:
        type: integer
      mediaURL:
        type: string
    type: object
  models.SectionReading:
    properties:
      mainText:
        type: string
    type: object
  models.SectionSpeaking:
    properties:
      audioUrl:
        type: string
    type: object
  models.SectionWriting:
    properties:
      minimumWords:
        type: integer
    type: object
  models.TestComponentStat:
    properties:
      count:
        type: number
      score:
        type: number
    type: object
...
  models.TestSkillsStat:
    properties:
      listening:
        items:
          $ref: '#/definitions/models.TestComponentStat'
        type: array
      overall:
        items:
          $ref: '#/definitions/models.TestComponentStat'
        type: array
      reading:
        items:
          $ref: '#/definitions/models.TestComponentStat'
        type: array
      speaking:
        items:
          $ref: '#/definitions/models.TestComponentStat'
        type: array
      writing:
        items:
          $ref: '#/definitions/models.TestComponentStat'
        type: array
    type: object
host: mel-api.go7.ir
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /section:
    post:
      consumes:
      - application/json
      description: Create Section
      parameters:
      - description: Create section
        in: body
        name: Section
        required: true
        schema:
          $ref: '#/definitions/DTO.SectionCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.SectionCreateResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Create Section
      tags:
      - Sections
    put:
      consumes:
      - application/json
      description: Update Section
      parameters:
      - description: Update test
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/DTO.SectionUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.SectionUpdateResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Update Section
      tags:
      - Sections
  /section/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Section by id
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Delete Section by id
      tags:
      - Sections
    get:
      consumes:
      - application/json
      description: Find Section by id
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.SectionResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Find Section by id
      tags:
      - Sections
  /section/all:
    get:
      consumes:
      - application/json
      description: Find all Sections
      parameters:
      - description: section ID
        in: query
        name: test_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DTO.SectionResponse'
            type: array
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Find All
      tags:
      - Sections
...
  /user:
    get:
      consumes:
      - application/json
      description: type of "data" field in response is array of "DTO.UserSignUpOrLoginUserResponse"
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: filter
        in: query
        name: filter
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: order by
        in: query
        name: order_by
        type: string
      - description: order direction
        in: query
        name: order_direction
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.PaginateResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Get all user with pagination
      tags:
      - Users
...
  /user-test:
    post:
      consumes:
      - application/json
      description: Create user-test
      parameters:
      - description: Create test
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/DTO.UserTestCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.UserTestCreateResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Create user-test
      tags:
      - User-tests
  /user-test/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user-test by id
      parameters:
      - description: user-test ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Delete user-test by id
      tags:
      - User-tests
    get:
      consumes:
      - application/json
      description: Find user-test by id
      parameters:
      - description: user-test ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.UserTestResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Find user-test by id
      tags:
      - User-tests
  /user-test/all:
    get:
      consumes:
      - application/json
      description: type of "data" field in response is array of "DTO.UserTestResponse"
      parameters:
      - description: user-test ID
        in: query
        name: id
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.PaginateResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Find all user's test with user id specification and pagination
      tags:
      - User-tests
  /user-test/submit:
    put:
      consumes:
      - application/json
      description: Submit Test
      parameters:
      - description: Test ID
        in: query
        name: test_id
        required: true
        type: string
      - description: User Test ID
        in: query
        name: user_test_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.UserTestResultResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Submit Test
      tags:
      - User-tests
  /user/auth:
    post:
      consumes:
      - application/json
      description: Login or signup user with firebase token
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.UserSignUpOrLoginUserResponse'
        "400":
          description: error message detail
      security:
      - ApiKeyAuth: []
      summary: Login or signup user
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"